บันทึกวิดีโอได้เเล้ว

+++ sic-spring/sic/src/main/java/com/softinter/sicapi/dto/request/SaveBusinessRequest.java
@@ -31,5 +31,5 @@ public class SaveBusinessRequest {
     private String fax;
     private UUID uploadGroupId;
     private boolean isActive;
-    private Long rowVersion;
+    private Integer rowVersion;
 }
