+++ sic-spring/sic/src/main/java/com/softinter/sicapi/dto/request/SaveBusinessRoleRequest.java
@@ -19,5 +19,5 @@ public class SaveBusinessRoleRequest {
     private Integer sortOrder;
     private boolean isActive = true;
     private EntityState state = EntityState.DETACHED;
-    private Long rowVersion;
+    private Integer rowVersion;
 }
